feat: patch isolation
