+++ src/index.ts
@@ -2,9 +2,6 @@ export { Heightmap } from './procgen/Heightmap'
 export { NoiseSampler } from './procgen/NoiseSampler'
 export { ProcLayer } from './procgen/ProcLayer'
 export { Biome, BlockType } from './procgen/Biome'
-export { PatchCache } from './procgen/PatchCache'
-export { PatchBaseCache } from './procgen/PatchBaseCache'
-export { PatchBlocksCache } from './procgen/PatchBlocksCache'
 
 // export type { MappingConf, MappingData, MappingRanges } from "./common/types"
 // export { DevHacks } from './tools/DevHacks'
